Add decimal precision to query duration measurements
- Changed ClickHouse schema: duration_ms from UInt32 to Float64
- Updated Event struct: DurationMS from int64 to float64
- Modified resolver to calculate duration with decimal precision (duration.Seconds() * 1000)
- Updated web UI to display duration with 2 decimal places
- Updated latency statistics to show 2 decimal places

This fixes the issue where very fast queries (<1ms) were showing '-'
instead of actual sub-millisecond durations like 0.25 ms, 0.50 ms, etc.

+++ db/clickhouse/init.sql
@@ -10,7 +10,7 @@ CREATE TABLE IF NOT EXISTS beyond_ads.dns_queries
     qclass LowCardinality(String),
     outcome LowCardinality(String),
     rcode LowCardinality(String),
-    duration_ms UInt32
+    duration_ms Float64
 )
 ENGINE = MergeTree
 ORDER BY (ts, qname)
